requirements.txt
<<<<<<< HEAD
# Data Science Libraries
#numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
matplotlib==3.4.2
seaborn==0.11.1
jupyterlab==3.1.7
tensorflow==2.5.0
torch==1.9.0
keras==2.4.3
scipy==1.7.0

# Natural Language Processing Libraries
nltk==3.6.2
gensim==4.0.1
spacy==3.1.0

# Additional Libraries
streamlit==0.82.0
openai==0.10.2
shelve==3.7.0
pylatex==1.4.1
markdown==3.3.4
fpdf==1.7.2
pptx==0.6.19
inflection==0.5.1
=======
streamlit
python-dotenv
openai
>>>>>>> 80ffc79cfed73c088a094f8c43ae950f5a6d9b34
